fix tests and todos
